separate examples and tests logic
